Effect Relative Placement and Color Generation Fix
Difficulty settings broke color generation for levels with more than
three tiles, so I re-did the random color generation algorithm. Also,
effects on collision were linked to box game objects, not the tiles, so
it looked like the effect was sliding off the screen. I changed it to
the tiles, so it should look like a proper effect.