Game Timing Change
Changed timing for box spawns, so players can see the next color a bit
earlier, making the game a bit easier. Sped up tile movement animation
for a snappier experience. Made the starting vertical velocity and the
ending velocity a bit faster so it isn't as boring in the beginning, and
more tense at max speed. Renamed TileGenerator to TileController as it
is more appropriate. Fixed Triangle prefab having an old Polygon2D
collider, making it detect collisions when the sprite doesn't actually
touch the tile.